chore: Rename azure_builder in azure_dashboard_builder
