+++ src/opex_dashboard/builder_factory.py
@@ -3,7 +3,7 @@ from typing import Any, Optional
 from opex_dashboard.resolver import OA3Resolver
 from opex_dashboard.error import InvalidBuilderError
 from opex_dashboard.builders.base import Builder
-from opex_dashboard.builders.azure_builder import AzDashboardBuilder
+from opex_dashboard.builders.azure_dashboard_builder import AzDashboardBuilder
 
 
 class BuilderFactory:
